refactor: separate generated 4cade catalog
